+++ src/mocks/handlers.ts
@@ -3,6 +3,8 @@ import myPageHandlers from './myPageHandlers';
 import pointHandlers from './pointHandlers';
 import recruitHandlers from './recruitHandlers';
 import studyMainHandlers from './studyMainHandlers';
+import timerHandlers from './timerHandlers';
+import todoHandlers from './todoHandlers';
 
 export const handlers = [
   ...recruitHandlers,
@@ -10,4 +12,6 @@ export const handlers = [
   ...pointHandlers,
   ...studyMainHandlers,
   ...kakaoHandler,
+  ...timerHandlers,
+  ...todoHandlers,
 ];
