+++ src/mocks/studyMainHandlers.ts
@@ -1,18 +1,9 @@
 import { http, HttpResponse } from 'msw';
-import { todoListData, dummyStudyList, dummyNotices, timersData } from './data/studyDetailMockData';
-import {
-  CreateTodoRequest,
-  UpdateStudyRequest,
-  EditTodoRequest,
-  ToggleTodoRequest,
-  NoticeRequest,
-  TimerRequest,
-} from '../types/request';
-import { TodoType } from '../types/interface';
+import { dummyStudyList, dummyNotices } from './data/studyDetailMockData';
+import { UpdateStudyRequest, NoticeRequest } from '../types/request';
 import { mockStudyRoomList } from './data/studyListMockData';
 
 const studyMainHandlers = [
-  // 스터디룸 목록 조회
   http.get(`/api/rooms`, ({ request }) => {
     console.log('스터디 목록 조회');
     const url = new URL(request.url);
@@ -57,109 +48,7 @@ const studyMainHandlers = [
       return new HttpResponse(JSON.stringify({ message: '잘못된 요청입니다.' }), { status: 400 });
     }
   }),
-  http.get('/todos/:studyId', ({ params }) => {
-    const studyId = Number(params.studyId);
-    const todoList = todoListData[studyId];
 
-    todoList.todos.forEach((todo) => {
-      if (todo.isRunning) {
-        todoList.studyTotalTime += 1;
-        todo.studyTime += 1;
-      }
-    });
-    if (todoList) return HttpResponse.json(todoList);
-    return HttpResponse.error();
-  }),
-  http.post('/todos/new/:studyId', async ({ params, request }) => {
-    try {
-      const studyId = Number(params.studyId);
-      const todoList = todoListData[studyId];
-      const body = (await request.json()) as CreateTodoRequest;
-
-      if (!body) {
-        return new HttpResponse(JSON.stringify({ message: '잘못된 요청 본문이거나 body가 없습니다.' }), {
-          status: 400,
-        });
-      }
-      const { todoName } = body;
-      const newTodo: TodoType = {
-        todoId: todoList.todos.length,
-        todoName: todoName as string,
-        isCompleted: false,
-        studyTime: 0,
-        isRunning: false,
-      };
-
-      todoList.todos.push(newTodo);
-
-      return new HttpResponse(
-        JSON.stringify({
-          message: '투두가 추가되었습니다.',
-        }),
-        { status: 200 },
-      );
-    } catch {
-      return new HttpResponse(JSON.stringify({ message: '잘못된 요청입니다.' }), { status: 400 });
-    }
-  }),
-  http.patch('/todos/change/studyId', async ({ params, request }) => {
-    try {
-      const studyId = Number(params.studyId);
-      const todoList = todoListData[studyId];
-      const body = (await request.json()) as EditTodoRequest;
-
-      if (!body || !todoList) {
-        return new HttpResponse(JSON.stringify({ message: '잘못된 요청 본문이거나 body가 없습니다.' }), {
-          status: 400,
-        });
-      }
-      const { todoName, todoId } = body;
-
-      const targetTodo = todoList.todos.find((todo) => todo.todoId === todoId);
-      if (!targetTodo) {
-        return new HttpResponse(JSON.stringify({ message: '존재하지 않는 투두입니다.' }), {
-          status: 404,
-        });
-      }
-
-      targetTodo.todoName = todoName;
-
-      return new HttpResponse(JSON.stringify({ message: '투두가 수정되었습니다.' }), {
-        status: 200,
-      });
-    } catch {
-      return new HttpResponse(JSON.stringify({ message: '잘못된 요청입니다.' }), { status: 400 });
-    }
-  }),
-  http.patch(`/todos/:todoId/complete`, async ({ params, request }) => {
-    try {
-      const todoId = Number(params.todoId);
-      const studyId = Number(params.studyId);
-      const body = (await request.json()) as ToggleTodoRequest;
-      const todoList = todoListData[studyId];
-      if (!body || !todoId || !studyId || !todoList) {
-        return new HttpResponse(JSON.stringify({ message: '잘못된 요청 본문이거나 body가 없습니다.' }), {
-          status: 400,
-        });
-      }
-      const { isCompleted } = body;
-
-      const targetTodo = todoList.todos.find((todo) => todo.todoId === todoId);
-      if (!targetTodo) {
-        return new HttpResponse(JSON.stringify({ message: '존재하지 않는 투두입니다.' }), {
-          status: 404,
-        });
-      }
-
-      targetTodo.isCompleted = isCompleted;
-
-      return new HttpResponse(JSON.stringify({ message: '투두가 수정되었습니다.' }), {
-        status: 200,
-      });
-    } catch {
-      return new HttpResponse(JSON.stringify({ message: '잘못된 요청입니다.' }), { status: 400 });
-    }
-  }),
   http.get(`/rooms/:studyId/notices`, ({ params }) => {
     const studyId = params.studyId;
     const notice = dummyNotices[Number(studyId)];
@@ -238,86 +127,6 @@ const studyMainHandlers = [
       return new HttpResponse(
         JSON.stringify({
           message: '공지가 삭제되었습니다.',
-        }),
-        { status: 200 },
-      );
-    } catch {
-      return new HttpResponse(JSON.stringify({ message: '잘못된 요청입니다.' }), { status: 400 });
-    }
-  }),
-  http.get('/timers/:studyId', async ({ params }) => {
-    try {
-      const studyId = Number(params.studyId);
-      const timers = timersData[studyId];
-      if (!studyId || !timers) {
-        return new HttpResponse(JSON.stringify({ message: '잘못된 요청입니다.' }), {
-          status: 400,
-        });
-      }
-      timers.forEach((timer) => {
-        if (timer.isRunning) {
-          timer.timerTime += 1;
-        }
-      });
-      return HttpResponse.json(timers);
-    } catch {
-      return new HttpResponse(JSON.stringify({ message: '잘못된 요청입니다.' }), { status: 400 });
-    }
-  }),
-  http.post('/timer/start/:studyId', async ({ params, request }) => {
-    try {
-      const studyId = Number(params.studyId);
-      const body = (await request.json()) as TimerRequest;
-
-      if (!studyId || !body) {
-        return new HttpResponse(JSON.stringify({ message: '잘못된 요청입니다.' }), {
-          status: 400,
-        });
-      }
-      const { userId, todoId } = body;
-      const userData = timersData[studyId].find((data) => data.userId === userId);
-      const todoData = todoListData[studyId].todos.find((todo) => todo.todoId === todoId);
-      if (!userData || !todoData) {
-        return new HttpResponse(JSON.stringify({ message: '잘못된 요청입니다.' }), {
-          status: 400,
-        });
-      }
-      userData.isRunning = true;
-      todoData.isRunning = true;
-
-      return new HttpResponse(
-        JSON.stringify({
-          message: '타이머가 시작 되었습니다.',
-        }),
-        { status: 200 },
-      );
-    } catch {
-      return new HttpResponse(JSON.stringify({ message: '잘못된 요청입니다.' }), { status: 400 });
-    }
-  }),
-  http.post('/timer/stop/:studyId', async ({ params, request }) => {
-    try {
-      const studyId = Number(params.studyId);
-      const body = (await request.json()) as TimerRequest;
-
-      if (!studyId || !body) {
-        return new HttpResponse(JSON.stringify({ message: '잘못된 요청입니다.' }), {
-          status: 400,
-        });
-      }
-      const { userId, todoId } = body;
-      const userData = timersData[studyId].find((data) => data.userId === userId);
-      const todoData = todoListData[studyId].todos.find((todo) => todo.todoId === todoId);
-      if (!userData || !todoData) {
-        return new HttpResponse(JSON.stringify({ message: '잘못된 요청입니다.' }), {
-          status: 400,
-        });
-      }
-      userData.isRunning = false;
-      todoData.isRunning = false;
-      return new HttpResponse(
-        JSON.stringify({
-          message: '타이머가 중지 되었습니다.',
         }),
         { status: 200 },
       );
